requirements.txt
numpy
pandas
tensorflow-gpu
sklearn
matplotlib
joblib
flask
flask-restful
gunicorn
